# fastapi/测试脚本/v2测试脚本/test/导出接口文档脚本/fastapi_complete_api_spec.json
{
...
  "openapi": "3.1.0",
  "info": {
    "title": "用户系统API",
    "description": "模块化的用户管理系统",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v2/share_system/create": {
      "post": {
        "tags": [
          "v2-share_system"
        ],
        "summary": "Create Share",
        "description": "创建文档分享链接",
        "operationId": "create_share_api_v2_share_system_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/share_system/my-shares": {
      "get": {
        "tags": [
          "v2-share_system"
        ],
        "summary": "Get My Shares",
        "description": "获取我的分享列表",
        "operationId": "get_my_shares_api_v2_share_system_my_shares_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/share_system/detail/{share_id}": {
      "get": {
        "tags": [
          "v2-share_system"
        ],
        "summary": "Get Share Detail",
        "description": "获取分享详情",
        "operationId": "get_share_detail_api_v2_share_system_detail__share_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "share_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Share Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/share_system/update/{share_id}": {
      "put": {
        "tags": [
          "v2-share_system"
        ],
        "summary": "Update Share",
        "description": "更新分享配置",
        "operationId": "update_share_api_v2_share_system_update__share_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "share_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Share Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShareRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/share_system/delete/{share_id}": {
      "delete": {
        "tags": [
          "v2-share_system"
        ],
        "summary": "Delete Share",
        "description": "删除分享",
        "operationId": "delete_share_api_v2_share_system_delete__share_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "share_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Share Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/share_system/public/{share_code}": {
      "post": {
        "tags": [
          "v2-share_system"
        ],
        "summary": "Access Shared Document",
        "description": "访问分享的文档（公开接口）",
        "operationId": "access_shared_document_api_v2_share_system_public__share_code__post",
        "parameters": [
          {
            "name": "share_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Share Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessShareRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "v2-share_system"
        ],
        "summary": "Get Shared Document",
        "description": "获取分享的文档（GET方式，用于直接链接访问）",
        "operationId": "get_shared_document_api_v2_share_system_public__share_code__get",
        "parameters": [
          {
            "name": "share_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Share Code"
            }
          },
          {
            "name": "password",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/share_system/download/{share_code}": {
      "get": {
        "tags": [
          "v2-share_system"
        ],
        "summary": "Download Shared Document",
        "description": "下载分享的文档",
        "operationId": "download_shared_document_api_v2_share_system_download__share_code__get",
        "parameters": [
          {
            "name": "share_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Share Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/share_system/stats": {
      "get": {
        "tags": [
          "v2-share_system"
        ],
        "summary": "Get Share Stats",
        "description": "获取分享统计",
        "operationId": "get_share_stats_api_v2_share_system_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/share_system/toggle-status/{share_id}": {
      "post": {
        "tags": [
          "v2-share_system"
        ],
        "summary": "Toggle Share Status",
        "description": "切换分享状态（启用/禁用）",
        "operationId": "toggle_share_status_api_v2_share_system_toggle_status__share_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "share_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Share Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccessShareRequest": {
        "properties": {
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password",
            "description": "分享密码（密码保护时必填）"
          }
        },
        "type": "object",
        "title": "AccessShareRequest"
      },
...
      "CreateShareRequest": {
        "properties": {
          "document_id": {
            "type": "integer",
            "title": "Document Id",
            "description": "文档ID"
          },
          "share_type": {
            "$ref": "#/components/schemas/ShareType",
            "description": "分享类型",
            "default": "public"
          },
          "share_password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Share Password",
            "description": "分享密码（密码保护时必填）"
          },
          "allow_download": {
            "type": "boolean",
            "title": "Allow Download",
            "description": "允许下载",
            "default": true
          },
          "allow_comment": {
            "type": "boolean",
            "title": "Allow Comment",
            "description": "允许评论",
            "default": true
          },
          "expire_hours": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expire Hours",
            "description": "过期时间（小时），NULL表示永不过期"
          }
        },
        "type": "object",
        "required": [
          "document_id"
        ],
        "title": "CreateShareRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublicDocumentResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "author_username": {
            "type": "string",
            "title": "Author Username"
          },
          "publish_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publish Time"
          },
          "view_count": {
            "type": "integer",
            "title": "View Count"
          },
          "allow_download": {
            "type": "boolean",
            "title": "Allow Download"
          },
          "allow_comment": {
            "type": "boolean",
            "title": "Allow Comment"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "content",
          "summary",
          "file_type",
          "file_size",
          "author_username",
          "publish_time",
          "view_count",
          "allow_download",
          "allow_comment"
        ],
        "title": "PublicDocumentResponse"
      },
...
      "ShareDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "document_id": {
            "type": "integer",
            "title": "Document Id"
          },
          "share_code": {
            "type": "string",
            "title": "Share Code"
          },
          "share_type": {
            "$ref": "#/components/schemas/ShareType"
          },
          "share_url": {
            "type": "string",
            "title": "Share Url"
          },
          "allow_download": {
            "type": "boolean",
            "title": "Allow Download"
          },
          "allow_comment": {
            "type": "boolean",
            "title": "Allow Comment"
          },
          "status": {
            "$ref": "#/components/schemas/ShareStatus"
          },
          "expire_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expire Time"
          },
          "view_count": {
            "type": "integer",
            "title": "View Count"
          },
          "download_count": {
            "type": "integer",
            "title": "Download Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "document_title": {
            "type": "string",
            "title": "Document Title"
          },
          "document_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Summary"
          },
          "today_views": {
            "type": "integer",
            "title": "Today Views"
          },
          "week_views": {
            "type": "integer",
            "title": "Week Views"
          },
          "month_views": {
            "type": "integer",
            "title": "Month Views"
          },
          "recent_access_logs": {
            "items": {
              "$ref": "#/components/schemas/AccessLogResponse"
            },
            "type": "array",
            "title": "Recent Access Logs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "share_code",
          "share_type",
          "share_url",
          "allow_download",
          "allow_comment",
          "status",
          "expire_time",
          "view_count",
          "download_count",
          "created_at",
          "updated_at",
          "document_title",
          "document_summary",
          "today_views",
          "week_views",
          "month_views",
          "recent_access_logs"
        ],
        "title": "ShareDetailResponse"
      },
      "ShareListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ShareResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "ShareListResponse"
      },
      "ShareResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "document_id": {
            "type": "integer",
            "title": "Document Id"
          },
          "share_code": {
            "type": "string",
            "title": "Share Code"
          },
          "share_type": {
            "$ref": "#/components/schemas/ShareType"
          },
          "share_url": {
            "type": "string",
            "title": "Share Url"
          },
          "allow_download": {
            "type": "boolean",
            "title": "Allow Download"
          },
          "allow_comment": {
            "type": "boolean",
            "title": "Allow Comment"
          },
          "status": {
            "$ref": "#/components/schemas/ShareStatus"
          },
          "expire_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expire Time"
          },
          "view_count": {
            "type": "integer",
            "title": "View Count"
          },
          "download_count": {
            "type": "integer",
            "title": "Download Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "document_title": {
            "type": "string",
            "title": "Document Title"
          },
          "document_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Summary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "share_code",
          "share_type",
          "share_url",
          "allow_download",
          "allow_comment",
          "status",
          "expire_time",
          "view_count",
          "download_count",
          "created_at",
          "updated_at",
          "document_title",
          "document_summary"
        ],
        "title": "ShareResponse"
      },
...
      "ShareStatus": {
        "type": "string",
        "enum": [
          "active",
          "expired",
          "disabled"
        ],
        "title": "ShareStatus"
      },
      "ShareType": {
        "type": "string",
        "enum": [
          "public",
          "private",
          "password"
        ],
        "title": "ShareType"
      },
...
      "UpdateShareRequest": {
        "properties": {
          "share_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShareType"
              },
              {
                "type": "null"
              }
            ],
            "description": "分享类型"
          },
          "share_password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Share Password",
            "description": "分享密码"
          },
          "allow_download": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allow Download",
            "description": "允许下载"
          },
          "allow_comment": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allow Comment",
            "description": "允许评论"
          },
          "expire_hours": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expire Hours",
            "description": "过期时间（小时）"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShareStatus"
              },
              {
                "type": "null"
              }
            ],
            "description": "分享状态"
          }
        },
        "type": "object",
        "title": "UpdateShareRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}